// src/app/model/stat.interface.ts
export interface Stat {
   id: number;
   playerId: number;
   player: string;
   teamId: number;
   team: string;
   points: number;
   rebounds: number;
   assists: number;
   steals: number;
   blocks: number;
   feildGoalsAttempted: number;
   feildGoalsMade: number;
   threePointersAttempted: number;
   threePointerMade: number;
   freeThrowsAttempted: number;
   freeThrowsMade: number;
 }